Moved signature buffer related code into offst-signature.
